requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
torch==2.10.0
torchvision==0.21.0
numpy==1.26.4
pandas==2.2.3
matplotlib==3.9.4
tqdm==4.67.1
Pillow==11.1.0
jupyter==1.1.1
wandb==0.19.9
requests==2.32.0
python-multipart==0.0.9
python-dotenv==1.0.0